+++ datatig/models/field_date.py
@@ -22,6 +22,15 @@ class FieldDateConfigModel(FieldConfigModel):
         v.set_value(obj.read(self._key))
         return v
 
+    def get_frictionless_csv_field_specifications(self):
+        return [
+            {
+                "name": "field_" + self.get_id(),
+                "title": self.get_title(),
+                "type": "date",
+            }
+        ]
+
 
 class FieldDateValueModel(FieldValueModel):
     def set_value(self, value):
@@ -29,3 +38,6 @@ class FieldDateValueModel(FieldValueModel):
 
     def get_value(self):
         return self._value
+
+    def get_frictionless_csv_data_values(self):
+        return [self._value]