+++ datatig/models/field_list_strings.py
@@ -18,6 +18,15 @@ class FieldListStringsConfigModel(FieldConfigModel):
         v.set_value(obj.read(self._key))
         return v
 
+    def get_frictionless_csv_field_specifications(self):
+        return [
+            {
+                "name": "field_" + self.get_id(),
+                "title": self.get_title(),
+                "type": "string",
+            }
+        ]
+
 
 class FieldListStringsValueModel(FieldValueModel):
     def set_value(self, value):
@@ -29,3 +38,9 @@ class FieldListStringsValueModel(FieldValueModel):
 
     def get_value(self):
         return self._value
+
+    def get_frictionless_csv_data_values(self):
+        if isinstance(self._value, list):
+            return [", ".join([str(i) for i in self._value])]
+        else:
+            return [self._value]