+++ datatig/models/field_string.py
@@ -21,6 +21,15 @@ class FieldStringConfigModel(FieldConfigModel):
         v.set_value(obj.read(self._key))
         return v
 
+    def get_frictionless_csv_field_specifications(self):
+        return [
+            {
+                "name": "field_" + self.get_id(),
+                "title": self.get_title(),
+                "type": "string",
+            }
+        ]
+
 
 class FieldStringValueModel(FieldValueModel):
     def set_value(self, value):
@@ -28,3 +37,6 @@ class FieldStringValueModel(FieldValueModel):
 
     def get_value(self):
         return self._value
+
+    def get_frictionless_csv_data_values(self):
+        return [self._value]