+++ datatig/models/field_url.py
@@ -22,6 +22,16 @@ class FieldURLConfigModel(FieldConfigModel):
         v.set_value(obj.read(self._key))
         return v
 
+    def get_frictionless_csv_field_specifications(self):
+        return [
+            {
+                "name": "field_" + self.get_id(),
+                "title": self.get_title(),
+                "type": "string",
+                "format": "uri",
+            }
+        ]
+
 
 class FieldURLValueModel(FieldValueModel):
     def set_value(self, value):
@@ -29,3 +39,6 @@ class FieldURLValueModel(FieldValueModel):
 
     def get_value(self):
         return self._value
+
+    def get_frictionless_csv_data_values(self):
+        return [self._value]